Fix #1 - Not respecting lineSpacing when set
There was an issue where if you set some custom attributes before
setting attributedText on the label, it wouldn't keep those properties
in the attributedString. Now when you set lineSpacing, and then set
attributedText on the label, it'll pull the properties already setup and
append them to the attributedString you're putting in.