Fix issue where setting attributedText could override properties (#10)
* Fix issue where setting attributedText could override properties

There was a regression in 0.0.5 that caused properties that were set on
the label to overwrite the properties of the attributedString the user
was setting on attributedText. This fixes that regression by reverting
the changes that were made to the didSet block of attributedText. Since
there's a valid use-case for wanting to set properties on the label
and have them overwrite the properties inside the attributedstring, I've
added another function `setAttributedText(_ attributedString:
NSAttributedString,
afterInheritingLabelAttributesAndConfiguringWithBlock block:
((NSMutableAttributedString) -> NSMutableAttributedString)?)` to support
that.

* Don't run swiftlint on carthage builds

* Add description for new setAttributedString function